- don't ask to remove files in font/ on upload

+++ bin/upload.py
@@ -68,12 +68,13 @@ for existing in existing_files:
 
 
 
-# delete all files on the server that don't exist locally
+# delete all files on the server that don't exist locally unless in /font
 #files_to_delete = set(key_map.keys()).difference(local_files)
 for key_name in set(key_map.keys()).difference(local_files):
-    response = raw_input("delete {}? ".format(key_name))
-    if response == 'y':
-        print "deleting."
-        key_map[key_name].delete()
+    if not key_name.startswith("font/"):
+        response = raw_input("delete {}? ".format(key_name))
+        if response == 'y':
+            print "deleting."
+            key_map[key_name].delete()
 
 print("done.")
